backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
llama-index==0.9.8
llama-index-llms-gemini==0.1.4
llama-index-embeddings-huggingface==0.1.4
llama-index-retrievers-bm25==0.1.3
pymupdf==1.23.8
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow>=10.2.0,<11.0.0
nest-asyncio==1.5.8
pandas==2.1.3
python-dotenv==1.0.0
matplotlib
ipython
